Move FrameCurry inside FrameOps.get
