+++ lib/ops.js
@@ -7,13 +7,29 @@ var __extends = (this && this.__extends) || function (d, b) {
 var frames_1 = require("./frames");
 var iterators_1 = require("./ops/iterators");
 exports.MetaMap = iterators_1.MetaMap;
+var FrameCurry = (function (_super) {
+    __extends(FrameCurry, _super);
+    function FrameCurry(_func, _source) {
+        _super.call(this);
+        this._func = _func;
+        this._source = _source;
+    }
+    FrameCurry.prototype.apply = function (argument, parameter) {
+        return this._func(this._source, argument);
+    };
+    return FrameCurry;
+}(frames_1.Frame));
+exports.FrameCurry = FrameCurry;
 var FrameOps = (function (_super) {
     __extends(FrameOps, _super);
     function FrameOps(context) {
         _super.call(this, context);
     }
     FrameOps.prototype.get = function (key, origin) {
-        return iterators_1.MetaMapExpr(origin);
+        return new frames_1.FrameExpr([
+            new FrameCurry(iterators_1.MetaMap, origin),
+            frames_1.FrameArg.here(),
+        ]);
     };
     return FrameOps;
 }(frames_1.Frame));
