+++ src/ops.ts
@@ -1,7 +1,19 @@
-import { Context, Frame, FrameArray, FrameExpr } from "./frames";
+import { Context, Frame, FrameArg, FrameArray, FrameExpr } from "./frames";
+import { MetaMap } from "./ops/iterators";
 
-import { MetaMap, MetaMapExpr } from "./ops/iterators";
+export interface ICurryFunction extends Function {
+  (source: Frame, block: Frame): Frame;
+}
 
+export class FrameCurry extends Frame {
+  constructor(protected _func: ICurryFunction, protected _source: Frame) {
+    super();
+  }
+
+  public apply(argument: Frame, parameter: Frame) {
+    return this._func(this._source, argument);
+  }
+}
 
 export class FrameOps extends Frame {
   constructor(context: Context) {
@@ -9,7 +21,10 @@ export class FrameOps extends Frame {
   }
 
   public get(key: string, origin: Frame): Frame {
-    return MetaMapExpr(origin);
+    return new FrameExpr([
+      new FrameCurry(MetaMap, origin),
+      FrameArg.here(),
+    ]);
   }
 }
